@@ -32,7 +32,7 @@
     "core-js": "^3.4.7",
     "css-loader": "^6.3.0",
     "html-webpack-plugin": "^5.3.2",
-    "jest": "^27.2.1",
+    "jest": "^29.3.1",
     "mini-css-extract-plugin": "^2.3.0",
     "sass": "^1.42.1",
     "sass-loader": "^12.1.0",
@@ -41,12 +41,13 @@
     "webpack-dev-server": "^4.2.1"
   },
   "dependencies": {
-    "axios": "^0.21.4",
+    "axios": "^1.2.2",
     "classnames": "^2.2.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.1.3",
-    "react-router-dom": "^5.1.2",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.6.1",
     "redux": "^4.0.4",
     "redux-logger": "^3.0.6"
   }
