@@ -1,73 +1,35 @@
-import React, { Component } from 'react'
-import { connect } from 'react-redux'
+import React, { useEffect, useState } from 'react'
+import { useSelector } from 'react-redux'
 import axios from 'axios'
-import { Link } from 'react-router-dom'
+import { Link, useParams } from 'react-router-dom'
 
-@connect(state => ({
-	versions: state.holybible.versions
-}), dispatch => ({}))
-class VerseList extends Component {
+export default function VerseList() {
+  const { vcode, bcode, chapter } = useParams()
+  const [verses, setVerses] = useState([])
+  const [loading, setLoading] = useState(true)
 
-	constructor(props, context) {
-		super(props, context)
-		this.state = {
-			vcode: null,
-			bcode: null,
-			chapter: 0,
-			verses: [],
-			fetching: false
-		}
-	}
-
-	componentDidMount() {
-		const { versions, match } = this.props
-		const { vcode, bcode, chapter } = match.params
-
-		this.updateVerses(vcode, bcode, chapter)
-	}
-
-	componentWillReceiveProps(newProps) {
-		const { versions, match } = newProps
-		const { vcode, bcode, chapter } = match.params
-
-		this.updateVerses(vcode, bcode, chapter)
-	}
-
-	updateVerses(newVcode, newBcode, newCnum) {
-		const { vcode, bcode, cnum } = this.state
-		if (vcode == newVcode && bcode == newBcode && cnum == newCnum) {
-			return
-		}
-
-		this.setState({
-			vcode: newVcode,
-			bcode: newBcode,
-			chapter: newCnum
-		})
-
-		axios.get(`/bible/${newVcode}/${newBcode}/${newCnum}.json`)
-			.then(r => this.setState({ verses: r.data }))
+  useEffect(() => {
+    setLoading(true)
+    axios.get(`/bible/${vcode}/${bcode}/${chapter}.json`)
+			.then(r => setVerses(r.data))
 			.catch(e => console.log(e))
-	}
+      .finally(() => {
+        setLoading(false)
+      })
+  }, [vcode, bcode, chapter])
 
-	render() {
-		const { vcode, bcode, chapter, verses, fetching } = this.state
+  if (loading) {
+    return <div className="loading">로딩 중...</div>
+  }
 
-		if (fetching) {
-			return <div className="loading">로딩 중...</div>
-		} else {
-			return (
-				<div className="verses">
-					<Link to={`/${vcode}/${bcode}`} className="btn_close btn_verse">&lt; 뒤로</Link>
-					<ol>
-						{verses.map((verse, index) =>
-							<li key={index}>{verse.content}</li>
-						)}
-					</ol>
-				</div>
-			)
-		}
-	}
+  return (
+    <div className="verses">
+      <Link to={`/${vcode}/${bcode}`} className="btn_close btn_verse">&lt; 뒤로</Link>
+      <ol>
+        {verses.map((verse, index) =>
+          <li key={index}>{verse.content}</li>
+        )}
+      </ol>
+    </div>
+  )
 }
-
-export default VerseList;
