@@ -1,30 +1,22 @@
-import React, { Component } from 'react'
-import PropTypes from 'prop-types'
-import { connect } from 'react-redux'
-import { Link } from 'react-router-dom'
+import React from 'react'
+import { useSelector } from 'react-redux'
+import { Link, useParams } from 'react-router-dom'
 
-@connect(state => ({
-	versions: state.holybible.versions
-}), dispatch => ({}))
-class BookList extends Component {
 
-	render() {
-		const { versions, match } = this.props
-		const { vcode } = match.params
+export default function BookList({}) {
+  const versions = useSelector(state => state.holybible.versions)
+  const { vcode } = useParams()
 
-		let version = versions.find(item => item.vcode === vcode)
+  let version = versions.find(item => item.vcode === vcode)
 
-		return (
-			<div>
-				<Link to={`/`} className="btn_close">&lt; 뒤로</Link>
-				<ul>
-					{ version.bibles.map(item =>
-						<li key={item.bcode}><Link to={`/${vcode}/${item.bcode}`}>{item.name}</Link></li>
-					)}
-				</ul>
-			</div>
-		)
-	}
+  return (
+    <div>
+      <Link to={`/`} className="btn_close">&lt; 뒤로</Link>
+      <ul>
+        { version.bibles.map(item =>
+          <li key={item.bcode}><Link to={`/${vcode}/${item.bcode}`}>{item.name}</Link></li>
+        )}
+      </ul>
+    </div>
+  )
 }
-
-export default BookList;
