@@ -1,8 +1,8 @@
-import React, { Component } from 'react'
+import React from 'react'
 import classnames from 'classnames'
-import { connect } from 'react-redux'
+import { useSelector } from 'react-redux'
 
-import { HashRouter, Route, Link } from 'react-router-dom'
+import { HashRouter, Routes, Route, Link } from 'react-router-dom'
 
 import VersionList from '../components/VersionList'
 import BookList from '../components/BookList'
@@ -10,52 +10,39 @@ import ChapterList from '../components/ChapterList'
 import VerseList from '../components/VerseList'
 import Today from '../components/Today'
 
-@connect(state => ({
-	versions: state.holybible.versions
-}), dispatch => ({}))
-class App extends Component {
-	constructor(props, context) {
-		super(props, context)
-		this.state = {
-			version: null,
-			book: null,
-			chapter: null,
-			verses: []
-		}
-	}
+export default function App() {
+  const versions = useSelector(state => state.holybible.versions)
 
-	render() {
-		const { versions } = this.props
+  if (versions.length == 0) {
+    return (
+      <div>로딩 중</div>
+    )
+  }
 
-		if (versions.length == 0) {
-			return (
-				<div>로딩 중</div>
-			)
-		}
-
-		return (
-			<HashRouter>
-				<div className={classnames('main', {'dark': false})}>
-					<header>
-						<h1 className="title"><Link to="/">Holybible</Link></h1>
-						<ul className="menu">
-              <li><a href="https://play.google.com/store/apps/details?id=net.joostory.holybible">Download</a></li>
-							<li><a href="https://oh-my-bible.tistory.com">About</a></li>
-						</ul>
-					</header>
-					<section>
-						<Route path="/" exact={true} component={VersionList} />
-						<Route path="/:vcode" exact={true} component={BookList} />
-						<Route path="/:vcode/:bcode" component={ChapterList} />
-					</section>
-					<article>
-						<Route path="/:vcode/:bcode/:chapter" component={VerseList} />
-						<Route path="*" component={Today} />
-					</article>
-				</div>
-			</HashRouter>
-		)
-	}
+  return (
+    <HashRouter>
+      <div className={classnames('main', {'dark': false})}>
+        <header>
+          <h1 className="title"><Link to="/">Holybible</Link></h1>
+          <ul className="menu">
+            <li><a href="https://play.google.com/store/apps/details?id=net.joostory.holybible">Download</a></li>
+            <li><a href="https://oh-my-bible.tistory.com">About</a></li>
+          </ul>
+        </header>
+        <section>
+          <Routes>
+            <Route path="/" exact={true} element={<VersionList />} />
+            <Route path="/:vcode" exact={true} element={<BookList />} />
+            <Route path="/:vcode/:bcode/*" element={<ChapterList />} />
+          </Routes>
+        </section>
+        <article>
+          <Routes>
+            <Route path="/:vcode/:bcode/:chapter" element={<VerseList />} />
+            <Route path="*" element={<Today />} />
+          </Routes>
+        </article>
+      </div>
+    </HashRouter>
+  )
 }
-
-export default App
